+++ src/replicate/message.rs
@@ -1,6 +1,6 @@
 /// A message sent over the network.
 #[derive(Debug, Clone, PartialEq)]
 pub struct Message {
-    start: usize,
-    length: Option<usize>,
+    start: u64,
+    length: Option<u64>,
 }
