+++ tests/model.rs
@@ -11,11 +11,11 @@ use rand::seq::SliceRandom;
 use rand::Rng;
 use std::u8;
 
-const MAX_FILE_SIZE: usize = 5 * 10; // 5mb
+const MAX_FILE_SIZE: u64 = 5 * 10; // 5mb
 
 #[derive(Clone, Debug)]
 enum Op {
-    Get { index: usize },
+    Get { index: u64 },
     Append { data: Vec<u8> },
     Verify,
 }
@@ -25,12 +25,12 @@ impl Arbitrary for Op {
         let choices = [0, 1, 2];
         match choices.choose(g).expect("Value should exist") {
             0 => {
-                let index: usize = g.gen_range(0, MAX_FILE_SIZE);
+                let index: u64 = g.gen_range(0, MAX_FILE_SIZE);
                 Op::Get { index }
             }
             1 => {
-                let length: usize = g.gen_range(0, MAX_FILE_SIZE / 3);
-                let mut data = Vec::with_capacity(length);
+                let length: u64 = g.gen_range(0, MAX_FILE_SIZE / 3);
+                let mut data = Vec::with_capacity(length as usize);
                 for _ in 0..length {
                     data.push(u8::arbitrary(g));
                 }
@@ -61,7 +61,7 @@ quickcheck! {
           if index >= insta.len() {
             assert_eq!(data, None);
           } else {
-            assert_eq!(data, Some(model[index].clone()));
+            assert_eq!(data, Some(model[index as usize].clone()));
           }
         },
         Op::Verify => {
